# sidecar/contracts/openapi-2.5.0.json
{
  "components": {
    "schemas": {
...
      "DeleteResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "status": {
            "const": "success",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "message"
        ],
        "title": "DeleteResponse",
        "type": "object"
      },
      "DocumentRecord": {
        "properties": {
          "extracted_markdown": {
            "title": "Extracted Markdown",
            "type": "string"
          },
          "file_path": {
            "title": "File Path",
            "type": "string"
          },
          "file_size": {
            "minimum": 0.0,
            "title": "File Size",
            "type": "integer"
          },
          "file_type": {
            "title": "File Type",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "ingested_at": {
            "title": "Ingested At",
            "type": "string"
          },
          "num_chunks": {
            "minimum": 0.0,
            "title": "Num Chunks",
            "type": "integer"
          },
          "num_pages": {
            "minimum": 0.0,
            "title": "Num Pages",
            "type": "integer"
          },
          "status": {
            "enum": [
              "indexed",
              "indexed_fallback"
            ],
            "title": "Status",
            "type": "string"
          },
          "used_fallback_embeddings": {
            "title": "Used Fallback Embeddings",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "filename",
          "file_path",
          "file_size",
          "num_pages",
          "num_chunks",
          "status",
          "ingested_at",
          "file_type",
          "used_fallback_embeddings",
          "extracted_markdown"
        ],
        "title": "DocumentRecord",
        "type": "object"
      },
...
      "ExportRequest": {
        "additionalProperties": false,
        "properties": {
          "export_format": {
            "default": "pdf",
            "enum": [
              "pdf",
              "docx",
              "html",
              "htm"
            ],
            "title": "Export Format",
            "type": "string"
          },
          "markdown_content": {
            "maxLength": 10000000,
            "minLength": 1,
            "title": "Markdown Content",
            "type": "string"
          }
        },
        "required": [
          "markdown_content"
        ],
        "title": "ExportRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IngestResponse": {
        "properties": {
          "extracted_markdown": {
            "title": "Extracted Markdown",
            "type": "string"
          },
          "file_size": {
            "title": "File Size",
            "type": "integer"
          },
          "file_type": {
            "title": "File Type",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "ingested_at": {
            "title": "Ingested At",
            "type": "string"
          },
          "num_chunks": {
            "title": "Num Chunks",
            "type": "integer"
          },
          "num_pages": {
            "title": "Num Pages",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "used_fallback_embeddings": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": false,
            "title": "Used Fallback Embeddings"
          }
        },
        "required": [
          "id",
          "filename",
          "file_size",
          "num_pages",
          "num_chunks",
          "extracted_markdown",
          "status",
          "ingested_at",
          "file_type"
        ],
        "title": "IngestResponse",
        "type": "object"
      },
      "LogDiagnosticQuery": {
        "additionalProperties": false,
        "properties": {
          "extra_paths": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 100,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra Paths"
          }
        },
        "title": "LogDiagnosticQuery",
        "type": "object"
      },
      "LogDiagnosticReportSchema": {
        "properties": {
          "anomalies": {
            "items": {
              "$ref": "#/components/schemas/AnomalyRecordSchema"
            },
            "title": "Anomalies",
            "type": "array"
          },
          "has_critical": {
            "title": "Has Critical",
            "type": "boolean"
          },
          "scanned_files": {
            "items": {
              "type": "string"
            },
            "title": "Scanned Files",
            "type": "array"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "total_lines_scanned": {
            "title": "Total Lines Scanned",
            "type": "integer"
          }
        },
        "required": [
          "scanned_files",
          "total_lines_scanned",
          "anomalies",
          "has_critical",
          "summary"
        ],
        "title": "LogDiagnosticReportSchema",
        "type": "object"
      },
      "PagePreviewResponse": {
        "properties": {
          "doc_id": {
            "title": "Doc Id",
            "type": "string"
          },
          "image_base64": {
            "title": "Image Base64",
            "type": "string"
          },
          "mime_type": {
            "default": "image/png",
            "title": "Mime Type",
            "type": "string"
          },
          "page_number": {
            "title": "Page Number",
            "type": "integer"
          },
          "total_pages": {
            "title": "Total Pages",
            "type": "integer"
          }
        },
        "required": [
          "doc_id",
          "page_number",
          "total_pages",
          "image_base64"
        ],
        "title": "PagePreviewResponse",
        "type": "object"
      },
...
      "SuccessResponse": {
        "properties": {
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "title": "SuccessResponse",
        "type": "object"
      },
...
      "UpdateDocumentRequest": {
        "additionalProperties": false,
        "properties": {
          "embedding_model": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "pattern": ".*\\S.*",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Model"
          },
          "markdown_content": {
            "maxLength": 10000000,
            "minLength": 1,
            "title": "Markdown Content",
            "type": "string"
          }
        },
        "required": [
          "markdown_content"
        ],
        "title": "UpdateDocumentRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OnlyRag V2 Python Sidecar Engine",
    "version": "2.5.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/agent/logs/analyze": {
      "post": {
        "description": "Scan OnlyRag V2 log files and return a structured anomaly diagnostic report.\nDetects: truncated JSON, VRAM thrashing, infinite tool-calling loops.",
        "operationId": "agent_logs_analyze_agent_logs_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogDiagnosticQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogDiagnosticReportSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Agent Logs Analyze"
      }
    },
...
    "/documents/{doc_id}": {
      "delete": {
        "operationId": "delete_document_documents__doc_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Document"
      },
      "get": {
        "operationId": "get_document_documents__doc_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Document"
      },
      "put": {
        "operationId": "update_document_documents__doc_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDocumentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Document"
      }
    },
    "/documents/{doc_id}/page-preview/{page_num}": {
      "get": {
        "operationId": "get_page_preview_documents__doc_id__page_preview__page_num__get",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "page_num",
            "required": true,
            "schema": {
              "title": "Page Num",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagePreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Page Preview"
      }
    },
    "/documents/{doc_id}/translate-inplace-stream": {
      "post": {
        "operationId": "translate_document_inplace_stream_endpoint_documents__doc_id__translate_inplace_stream_post",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranslateInplaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Translate Document Inplace Stream Endpoint"
      }
    },
    "/export": {
      "post": {
        "operationId": "export_document_export_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Document"
      }
    },
...
    "/history/index": {
      "post": {
        "operationId": "index_history_history_index_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndexPromptHistoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Index History"
      }
    },
    "/history/remove": {
      "post": {
        "operationId": "remove_history_history_remove_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptHistoryRemoveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove History"
      }
    },
    "/history/search": {
      "post": {
        "operationId": "search_history_history_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptHistorySearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PromptHistorySearchResult"
                  },
                  "title": "Response Search History History Search Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search History"
      }
    },
    "/ingest-path-stream": {
      "post": {
        "operationId": "ingest_document_by_path_stream_ingest_path_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestPathRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Document By Path Stream"
      }
    },
    "/tasks/cancel": {
      "post": {
        "operationId": "cancel_sidecar_task_tasks_cancel_post",
        "parameters": [
          {
            "in": "query",
            "name": "task_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCancelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Sidecar Task"
      }
    },
    "/vector/search": {
      "post": {
        "operationId": "search_vector_db_vector_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SearchResult"
                  },
                  "title": "Response Search Vector Db Vector Search Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Vector Db"
      }
    }
  }
}